Update to ppmck 9a ex11.3 am1, fork by AoiMoe

+++ src/ppmckc/file.c
@@ -99,7 +99,7 @@ int getFileSize( char *ptr )
 	if( fp == NULL ) return 0;
 
 	fseek( fp, 0L, SEEK_END );
-	size = ftell( fp );
+	size = (int)ftell( fp );
 	fseek( fp, 0L, SEEK_SET );
 	fclose( fp );
 
@@ -117,7 +117,7 @@ int getFileSize( char *ptr )
 --------------------------------------------------------------*/
 char *readTextFile( char *filename )
 {
-
+	
 	FILE *fp;
 	int size, sizeb, line_idx;
 	char *top, *p;
@@ -134,7 +134,7 @@ char *readTextFile( char *filename )
 	}
 	/* サイズを取得 */
 	fseek(fp, 0L, SEEK_END);
-	sizeb = ftell(fp);
+	sizeb = (int)ftell(fp);
 	fseek(fp, 0L, SEEK_SET);
 	/*
 	if (sizeb == 0) {
@@ -195,7 +195,7 @@ char *readTextFile( char *filename )
 					free(top);
 					printf("%s : ungetc failed\n", filename);
 					return NULL;
-
+					
 				}
 			}
 		} else if (c == '\x0a'){
@@ -218,7 +218,7 @@ char *readTextFile( char *filename )
 		}
 	}
 	fclose(fp);
-	size = (p - top) / sizeof(c);
+	size = (int)(p - top) / sizeof(c);
 	/*
 	printf("read %d byte -> store %d byte (\\0 を含む) \n", sizeb, size);
 	printf("read %d line\n", line_idx);
@@ -253,14 +253,14 @@ static void initDmcPath(void)
 	p = getenv("DMC_INCLUDE");
 	if (p == NULL)
 		return;
-
+	
 	for (i = 0; i < 10; i++) {
 		pl = strchr(p, ';');
 
 		if (pl == NULL)
-			l = strlen(p);
+			l = (int)strlen(p);
 		else
-			l = pl-p;
+			l = (int)(pl-p);
 
 		if (l == 0) {
 			dmcpath[i][0] = '\0';
@@ -287,7 +287,7 @@ FILE *openDmc(char *name)
 	if (!dmcpath_inited) {
 		initDmcPath();
 	}
-
+	
 	fileptr = fopen(name, "rb");
 	if (fileptr != NULL) return(fileptr);
 
@@ -295,7 +295,7 @@ FILE *openDmc(char *name)
 		if (strlen(dmcpath[i])) {
 			strcpy(testname, dmcpath[i]);
 			strcat(testname, name);
-
+	
 			fileptr = fopen(testname, "rb");
 			if (fileptr != NULL) break;
 		}
